Move config file to backend root
