+++ src/datadoc/backend/src/config.py
@@ -27,7 +27,6 @@ DAPLA_SERVICE = "DAPLA_SERVICE"
 env_loaded = False
 
 
-# TODO(@tilen1976): necessary for backend?  # noqa: TD003
 def _load_dotenv_file() -> None:
     global env_loaded  # noqa: PLW0603
     if not env_loaded and DOT_ENV_FILE_PATH.exists():
@@ -52,13 +51,11 @@ def get_jupyterhub_user() -> str | None:
     return _get_config_item(JUPYTERHUB_USER)
 
 
-# TODO(@tilen1976): necessary for backend?  # noqa: TD003
 def get_datadoc_dataset_path() -> str | None:
     """Get the path to the dataset."""
     return _get_config_item("DATADOC_DATASET_PATH")
 
 
-# TODO(@tilen1976): add logger config?  # noqa: TD003
 def get_log_level() -> int:
     """Get the log level."""
     # Magic numbers as defined in Python's stdlib logging
@@ -78,7 +75,6 @@ def get_log_level() -> int:
         return log_levels["INFO"]
 
 
-# TODO(@tilen1976): add logger config?  # noqa: TD003
 def get_log_formatter() -> Literal["simple", "json"]:
     """Get log formatter configuration."""
     if (
@@ -145,25 +141,21 @@ def get_oidc_token() -> str | None:
     return _get_config_item("OIDC_TOKEN")
 
 
-# TODO(@tilen1976): remove backend?  # noqa: TD003
 def get_unit_code() -> int | None:
     """The code for the Unit Type code list in Klass."""
     return int(_get_config_item("DATADOC_UNIT_CODE") or 702)
 
 
-# TODO(@tilen1976): remove backend?  # noqa: TD003
 def get_measurement_unit_code() -> int | None:
     """The code for the Measurement Unit code list in Klass."""
     return int(_get_config_item("DATADOC_MEASUREMENT_UNIT") or 303)
 
 
-# TODO(@tilen1976): remove backend?  # noqa: TD003
 def get_organisational_unit_code() -> int | None:
     """The code for the organisational units code list in Klass."""
     return int(_get_config_item("DATADOC_ORGANISATIONAL_UNIT_CODE") or 83)
 
 
-# TODO(@tilen1976): remove backend?  # noqa: TD003
 def get_data_source_code() -> int | None:
     """The code for the organisational units code list in Klass."""
     return int(_get_config_item("DATADOC_DATA_SOURCE_CODE") or 712)
